requirements.txt
streamlit==1.51.0
fastapi==0.121.1
uvicorn==0.38.0
requests==2.32.5
pillow==12.0.0
python-dotenv==1.2.1
pdf2image==1.17.0
python-multipart==0.0.20